lol: wallet: use floating point calculation for average output time, adapted from #ba0d55f2

+++ src/wallet/logic/state/gamma_picker.cpp
@@ -48,12 +48,15 @@ namespace state {
     THROW_WALLET_EXCEPTION_IF(rct_offsets.size() <= CRYPTONOTE_DEFAULT_TX_SPENDABLE_AGE, tools::error::wallet_internal_error, "Bad offset calculation");
     const size_t blocks_in_a_year = 86400 * 365 / constant::DIFFICULTY_TARGET_IN_SECONDS;
     const size_t blocks_to_consider = std::min<size_t>(rct_offsets.size(), blocks_in_a_year);
-    const size_t outputs_to_consider = rct_offsets.back() - (blocks_to_consider < rct_offsets.size() ? rct_offsets[rct_offsets.size() - blocks_to_consider - 1] : 0);
+    const size_t outputs_to_consider = rct_offsets.back() -
+      (blocks_to_consider < rct_offsets.size() ? rct_offsets[rct_offsets.size() - blocks_to_consider - 1] : 0);
     begin = rct_offsets.data();
     end = rct_offsets.data() + rct_offsets.size() - CRYPTONOTE_DEFAULT_TX_SPENDABLE_AGE;
     num_rct_outputs = *(end - 1);
     THROW_WALLET_EXCEPTION_IF(num_rct_outputs == 0, tools::error::wallet_internal_error, "No rct outputs");
-    average_output_time = constant::DIFFICULTY_TARGET_IN_SECONDS * blocks_to_consider / outputs_to_consider; // this assumes constant target over the whole rct range
+
+    average_output_time = constant::DIFFICULTY_TARGET_IN_SECONDS * blocks_to_consider /
+      static_cast<double>(outputs_to_consider); // this assumes constant target over the whole rct range
   };
 
   gamma_picker::gamma_picker(const std::vector<uint64_t> &rct_offsets): gamma_picker(rct_offsets, GAMMA_SHAPE, GAMMA_SCALE) {}
